Move ResourceBundle Files into Packages
Moved the ResourceBundle files into a package so there will be no more
naming collisions if the files are named equally in a project which uses
this library. The ResourceBundle for translation is now found by using
the fully qualified name of the bundle instead of (non unique) "simple"
name.